+++ model_angl/cell.hpp
@@ -7,15 +7,15 @@
 
 class Cell {
 public:
-    Cell(int x, int y, int state) {
+    Cell(int x, int y, double state) {
         setCell(x,y,state);
     }
     Cell(void) {
         setCell(0,0,0);
     }
-    void updateDensity(int state);
-    void setCell(int x, int y, int state);
-    int state;
+    void updateDensity(double state);
+    void setCell(int x, int y, double state);
+    double state;
     int x;
     int y;
 };
